resultados realizados com o script python
