--- MPI/tempo_par
@@ -1,47 +1,0 @@
-8
-1000 0.033238
-334300 0.088207
-667600 0.154885
-1000900 0.217870
-1334200 0.266914
-1667500 0.323678
-2000800 0.414084
-2334100 0.465860
-2667400 0.519843
-3000700 0.587615
-3334000 0.681011
-3667300 0.724931
-4000600 0.751814
-4333900 0.811338
-4667200 0.943082
-5000500 1.065962
-5333800 1.044554
-5667100 1.088479
-6000400 1.171306
-6333700 1.216931
-6667000 1.329198
-7000300 1.397521
-7333600 1.402366
-7666900 1.460687
-8000200 1.603686
-8333500 1.611440
-8666800 1.646158
-9000100 1.777401
-9333400 1.873007
-9666700 1.831300
-10000000 1.907322
-20 0.036539
-1000 0.038002
-20 0.034153
-20 0.036819
-20 0.044609
-20 0.046397
-20 0.031626
-20 0.036106
-20 0.079991
-20 0.084075
-20 0.045360
-20 0.042476
-20 0.043009
-20 0.101654
-20 0.098281
