--- openMP/entrada1
@@ -1,1001 +1,0 @@
-1000
-6074
-6819
-3267
-2751
-2439
-2287
-6995
-5441
-490
-85
-9343
-8589
-7547
-3502
-4367
-9320
-2747
-5096
-2869
-3325
-4086
-1407
-1781
-3893
-4594
-8495
-2630
-7489
-9493
-1911
-2158
-5567
-5083
-5425
-4670
-7522
-7713
-8018
-2963
-4555
-8103
-8658
-3144
-5650
-8512
-3864
-1322
-7611
-5312
-4191
-7288
-9398
-5599
-9069
-3292
-193
-7564
-5922
-4034
-3410
-4185
-6192
-5329
-9268
-7970
-0
-3142
-2035
-4370
-2457
-6590
-2473
-1115
-6086
-8123
-5979
-6302
-5797
-3591
-1615
-9989
-879
-1013
-5588
-9949
-657
-2133
-3865
-2931
-2519
-7275
-7117
-8712
-2605
-6385
-3034
-8957
-9528
-5069
-3327
-1985
-1659
-5800
-9453
-4097
-275
-5432
-400
-6072
-9023
-2015
-2413
-6255
-9380
-4353
-6204
-38
-6486
-69
-9321
-9006
-3697
-6438
-4070
-6302
-9176
-3456
-1611
-5056
-8525
-4938
-7041
-6536
-7090
-2846
-633
-7365
-8279
-7385
-9789
-3654
-5752
-2203
-9909
-5133
-6556
-2465
-1523
-9395
-8887
-7196
-4753
-8936
-3635
-5175
-5238
-9163
-8631
-6849
-4219
-7156
-8139
-7612
-44
-5229
-459
-677
-8946
-5090
-4415
-8735
-8744
-167
-7290
-5006
-1652
-199
-3823
-9527
-9594
-9062
-3076
-699
-7998
-3063
-5874
-9588
-2226
-4505
-2789
-6445
-8013
-928
-4057
-8057
-2509
-4516
-5086
-1455
-5958
-9501
-6543
-4703
-6021
-3833
-6061
-4025
-384
-6236
-9905
-6330
-5299
-2981
-7029
-9649
-6044
-2903
-9238
-8270
-3760
-2027
-1067
-1773
-9308
-5124
-9830
-1817
-5993
-4917
-3273
-1951
-770
-9816
-3006
-6791
-1
-5419
-817
-6738
-1656
-7074
-3068
-3307
-55
-6450
-9308
-6099
-5705
-8546
-721
-9466
-6926
-1788
-1239
-6234
-6912
-7422
-8051
-9257
-2339
-1324
-1209
-3109
-7492
-567
-6253
-3846
-5987
-3422
-584
-3995
-496
-4
-3654
-551
-6454
-2962
-3002
-2160
-7861
-3723
-1626
-4787
-1863
-9217
-1021
-8775
-6639
-9072
-4385
-5330
-6749
-1946
-8440
-4241
-2513
-1045
-8087
-4852
-4467
-5023
-5199
-1315
-5028
-5205
-8218
-1482
-4520
-7572
-3642
-2381
-1295
-1620
-7168
-3158
-7190
-4541
-8285
-3829
-3613
-2670
-9160
-6714
-968
-3952
-956
-3482
-4997
-5395
-4686
-5816
-419
-9886
-3483
-5447
-1443
-1701
-6929
-5963
-9273
-6924
-8344
-6920
-8544
-1864
-6430
-5734
-2757
-4715
-5916
-6371
-3738
-5076
-9437
-1058
-5380
-393
-4540
-6729
-2141
-5579
-2545
-2560
-1817
-6028
-8007
-3260
-4081
-1288
-9224
-3354
-8212
-3920
-274
-3109
-2137
-3056
-5195
-4894
-4123
-7463
-7617
-7861
-8891
-7055
-8920
-4271
-3800
-9812
-1000
-5941
-5391
-9897
-8501
-7208
-5925
-2860
-469
-6
-4149
-6045
-9712
-8713
-9965
-6338
-8174
-8454
-9394
-9722
-3349
-3518
-7185
-966
-7731
-6077
-4373
-3003
-348
-8174
-2816
-7701
-4115
-4559
-7598
-8969
-8120
-9876
-1829
-4941
-6234
-2330
-7338
-5947
-1044
-3655
-8637
-5570
-2110
-8032
-5292
-5459
-7902
-2478
-2777
-5633
-4907
-7151
-8637
-1607
-5325
-7805
-9308
-5792
-8716
-3259
-4761
-6836
-3135
-2943
-8129
-9369
-5273
-5467
-1668
-2669
-9123
-306
-4592
-1233
-4690
-9884
-3044
-2592
-8714
-5821
-4577
-3621
-9324
-3214
-5229
-1001
-7371
-889
-6794
-6088
-4148
-7907
-9276
-3635
-850
-7406
-9357
-2476
-9225
-1025
-1497
-8348
-7683
-6089
-9581
-8725
-5974
-8977
-7669
-4688
-4799
-2247
-4662
-475
-1813
-9891
-1477
-9185
-7132
-4623
-1625
-1281
-2530
-901
-4916
-9733
-4659
-4273
-8561
-3885
-1651
-58
-8585
-5686
-2500
-4519
-764
-4826
-3496
-8433
-5866
-8295
-680
-528
-5123
-2494
-6771
-6600
-8031
-3904
-1223
-9656
-1537
-105
-6909
-6453
-6190
-7921
-7079
-4751
-8158
-5082
-1162
-6743
-7120
-14
-1262
-7884
-4840
-1111
-6318
-706
-5758
-3350
-7587
-881
-2196
-4358
-7481
-227
-4614
-5056
-9883
-6151
-5162
-3145
-8957
-7704
-1066
-2388
-8808
-5576
-7470
-6322
-2319
-942
-6336
-9934
-8827
-1176
-1045
-1497
-8234
-6803
-1199
-5821
-7685
-3396
-180
-1518
-3623
-4794
-6575
-9859
-7298
-8089
-9356
-6255
-5793
-422
-8643
-953
-2350
-2465
-7275
-4669
-3407
-3611
-955
-8586
-1139
-2000
-6435
-5726
-5156
-7635
-1547
-9193
-1031
-8079
-711
-4654
-9226
-3638
-865
-6524
-1727
-221
-9131
-3873
-6995
-4126
-4826
-9345
-6591
-8454
-367
-6350
-8417
-1322
-1289
-9557
-9675
-7724
-5283
-4831
-5359
-3182
-376
-2742
-1262
-1087
-7397
-488
-4726
-4614
-3364
-2805
-4836
-2495
-6678
-1831
-6621
-7857
-7529
-9564
-6311
-7896
-2266
-4728
-9218
-3555
-637
-5245
-7632
-2272
-6428
-2991
-5455
-6804
-5734
-3069
-7892
-9483
-3557
-8970
-4097
-6921
-1775
-5285
-5768
-4806
-3469
-2389
-2663
-998
-8305
-5326
-8894
-571
-6406
-4464
-4127
-7044
-6062
-1759
-5668
-2490
-1102
-7475
-9295
-3188
-544
-3539
-9023
-4101
-8861
-9473
-7374
-636
-4758
-9494
-1794
-8227
-8235
-809
-5577
-6540
-6135
-823
-7112
-8894
-1640
-7591
-2290
-7702
-5702
-7958
-192
-3156
-5434
-9487
-6345
-5978
-9378
-5368
-6432
-8239
-1193
-158
-5228
-5952
-9653
-7022
-531
-4240
-7832
-2461
-781
-319
-3284
-4245
-5565
-4924
-8188
-7855
-2626
-242
-5814
-9171
-3398
-7600
-5010
-6095
-9930
-4389
-7816
-2714
-8980
-9009
-2873
-4208
-4961
-8878
-1231
-1845
-3118
-5415
-4306
-251
-2086
-7590
-848
-7652
-8867
-9036
-5507
-7845
-9278
-7673
-7016
-9029
-5273
-8379
-5124
-1556
-9120
-9292
-4270
-8100
-8302
-3495
-2309
-9615
-2373
-9892
-1460
-1844
-1659
-2118
-8447
-97
-6061
-9296
-7749
-1280
-4684
-9609
-9125
-3963
-7282
-2494
-9344
-8908
-873
-820
-6816
-6345
-113
-1086
-4445
-8415
-4582
-3106
-4382
-3307
-2998
-2195
-5151
-1009
-4313
-3599
-1107
-6726
-9247
-5208
-8006
-3931
-4817
-7132
-4246
-8452
-5978
-3590
-3712
-3203
-4411
-528
-9548
-4524
-7966
-3993
-9291
-2548
-3452
-3673
-5856
-2802
-2220
-7359
-3812
-6534
-7310
-4919
-3260
-6557
-6479
-1267
-489
-1297
-4751
-4735
-6101
-729
-4678
-9813
-284
-9089
-6693
-9832
-9965
-4659
-177
-9256
-7208
-3629
-9281
-9416
-6432
-1502
-3127
-244
-4388
-438
-1515
-7648
-6995
-7994
-5267
-3836
-5643
-6370
-4924
-1744
-7099
-9602
-1557
-7383
-5043
-4602
-3567
-5008
-9262
-3745
-4264
-2822
-7374
-9897
-2238
-158
-1399
-5365
-6754
-5787
-2155
-8269
-9788
-5503
-2616
-1407
-9339
-4611
-7778
-615
-6356
-1229
-6569
-4265
-8613
-1612
-8868
-8532
-6620
-4482
-2277
-7236
-7304
-6004
-7134
-5894
-6162
-4885
-1259
-2917
-7025
-9767
-7538
-6813
-5270
-6506
-8220
-961
-1118
-2350
-1577
