+++ openMP/tempo_par
@@ -1,0 +1,1 @@
+10000000 0.546113
