--- openMP/tempo_seq
@@ -1,102 +1,0 @@
-20 0.000332
-20 0.000260
-20 0.000110
-20 0.000098
-20 0.000000
-20 0.000000
-20 0.000644
-20 0.000615
-20 0.000553
-20 0.000638
-20 0.000575
-20 0.000589
-20 0.000726
-20 0.000510
-20 0.000626
-20 0.000588
-20 0.000441
-20 0.000645
-20 0.000544
-20 0.000438
-20 0.000176
-20 0.000999
-20 0.001283
-20 0.000522
-20 0.000639
-20 0.001235
-20 0.001187
-20 0.000712
-20 0.000741
-20 0.000836
-20 0.000844
-20 0.001157
-20 0.001216
-20 0.003096
-20 0.001187
-20 0.001280
-20 0.000428
-20 0.000415
-20 0.001253
-20 0.001130
-20 0.001126
-20 0.001211
-20 0.001269
-20 0.001395
-20 0.000499
-20 0.000368
-20 0.000258
-20 0.001302
-20 0.000576
-20 0.001393
-20 0.001326
-20 0.000581
-20 0.001097
-20 0.001207
-20 0.004754
-20 0.000195
-20 0.001296
-20 0.000361
-20 0.001184
-20 0.001234
-20 0.001357
-20 0.000269
-20 0.001304
-20 0.000282
-20 0.001424
-20 0.001267
-20 0.001208
-20 0.000340
-20 0.001293
-20 0.000357
-20 0.000356
-20 0.000372
-20 0.000352
-20 0.001099
-20 0.001192
-20 0.001240
-20 0.000404
-20 0.000354
-20 0.001329
-20 0.001294
-20 0.001337
-20 0.001187
-20 0.001513
-20 0.001442
-20 0.001350
-20 0.001191
-20 0.000495
-20 0.000319
-20 0.000295
-20 0.001248
-20 0.001375
-20 0.001390
-20 0.000587
-20 0.001290
-20 0.000422
-20 0.000271
-20 0.000354
-20 0.001277
-20 0.000360
-20 0.000372
-20 0.001332
-20 0.001430
