+++ resultados/tempo_par_2
@@ -1,0 +1,31 @@
+1000 0,033852
+334300 0,091530
+667600 0,125880
+1000900 0,170359
+1334200 0,195386
+1667500 0,239089
+2000800 0,271850
+2334100 0,313072
+2667400 0,336510
+3000700 0,408493
+3334000 0,408910
+3667300 0,485553
+4000600 0,489135
+4333900 0,530806
+4667200 0,567667
+5000500 0,622227
+5333800 0,675457
+5667100 0,752873
+6000400 0,756237
+6333700 0,801156
+6667000 0,817874
+7000300 0,892792
+7333600 0,905862
+7666900 0,964102
+8000200 0,991805
+8333500 1,008568
+8666800 1,047207
+9000100 1,137116
+9333400 1,113833
+9666700 1,217172
+10000000 1,234076
