+++ resultados/tempo_par_4
@@ -1,0 +1,31 @@
+1000 0,029946
+334300 0,068361
+667600 0,086637
+1000900 0,116394
+1334200 0,138321
+1667500 0,146485
+2000800 0,180971
+2334100 0,179575
+2667400 0,207211
+3000700 0,240906
+3334000 0,258016
+3667300 0,268361
+4000600 0,319062
+4333900 0,335239
+4667200 0,369604
+5000500 0,382309
+5333800 0,393118
+5667100 0,442608
+6000400 0,428235
+6333700 0,472987
+6667000 0,487850
+7000300 0,516424
+7333600 0,529860
+7666900 0,577583
+8000200 0,595603
+8333500 0,614946
+8666800 0,639894
+9000100 0,643553
+9333400 0,679815
+9666700 0,753295
+10000000 0,778516
