+++ resultados/tempo_par_8
@@ -1,0 +1,31 @@
+1000 0,073184
+334300 0,094713
+667600 0,106327
+1000900 0,119819
+1334200 0,136782
+1667500 0,165732
+2000800 0,173204
+2334100 0,196523
+2667400 0,221320
+3000700 0,242905
+3334000 0,258412
+3667300 0,273819
+4000600 0,293098
+4333900 0,322668
+4667200 0,343450
+5000500 0,359299
+5333800 0,393694
+5667100 0,399792
+6000400 0,442082
+6333700 0,432849
+6667000 0,437883
+7000300 0,497123
+7333600 0,493793
+7666900 0,516340
+8000200 0,560970
+8333500 0,597922
+8666800 0,578016
+9000100 0,598831
+9333400 0,623777
+9666700 0,643708
+10000000 0,703037
