+++ resultados/tempo_par_omp_2
@@ -1,0 +1,31 @@
+1000 0,000169
+334300 0,039004
+667600 0,078429
+1000900 0,105182
+1334200 0,154572
+1667500 0,191828
+2000800 0,230089
+2334100 0,261274
+2667400 0,299029
+3000700 0,351974
+3334000 0,389757
+3667300 0,432174
+4000600 0,471826
+4333900 0,504158
+4667200 0,567609
+5000500 0,605319
+5333800 0,643984
+5667100 0,708989
+6000400 0,733836
+6333700 0,745785
+6667000 0,761747
+7000300 0,782568
+7333600 0,81404
+7666900 0,8762
+8000200 0,893514
+8333500 0,964102
+8666800 0,995438
+9000100 1,028606
+9333400 1,068376
+9666700 1,107366
+10000000 1,148291
