+++ resultados/tempo_par_omp_4
@@ -1,0 +1,31 @@
+1000 0,000225
+334300 0,033344
+667600 0,041958
+1000900 0,070757
+1334200 0,079924
+1667500 0,102036
+2000800 0,121129
+2334100 0,142879
+2667400 0,177988
+3000700 0,194777
+3334000 0,207958
+3667300 0,231986
+4000600 0,251076
+4333900 0,27144
+4667200 0,295197
+5000500 0,317977
+5333800 0,364114
+5667100 0,374275
+6000400 0,401719
+6333700 0,424144
+6667000 0,453403
+7000300 0,468516
+7333600 0,4805
+7666900 0,5158
+8000200 0,551569
+8333500 0,567061
+8666800 0,583371
+9000100 0,614484
+9333400 0,652376
+9666700 0,65182
+10000000 0,675683
