+++ resultados/tempo_par_omp_8
@@ -1,0 +1,31 @@
+1000 0,014531
+334300 0,03304
+667600 0,045388
+1000900 0,059103
+1334200 0,071345
+1667500 0,089515
+2000800 0,111977
+2334100 0,136963
+2667400 0,14495
+3000700 0,161433
+3334000 0,181163
+3667300 0,198727
+4000600 0,215845
+4333900 0,235315
+4667200 0,252615
+5000500 0,275157
+5333800 0,292809
+5667100 0,311828
+6000400 0,335777
+6333700 0,347277
+6667000 0,377116
+7000300 0,382829
+7333600 0,398503
+7666900 0,424119
+8000200 0,446565
+8333500 0,453331
+8666800 0,472063
+9000100 0,511368
+9333400 0,541296
+9666700 0,536259
+10000000 0,546912
