--- resultados/tempo_par_proc
@@ -1,9 +1,0 @@
-10000000 2.646967 2
-10000000 2.579867 2
-10000000 2.545378 2
-10000000 1.974478 4
-10000000 2.049065 4
-10000000 1.994688 4
-10000000 1.853637 8
-10000000 1.879759 8
-10000000 1.941172 8
