--- resultados/tempo_par_proc_media
@@ -1,3 +1,0 @@
-2 2.590737
-4 2.006077
-8 1.891523
