--- resultados/tempo_seq
@@ -1,31 +1,0 @@
-1000 0.000598
-334300 0.066522
-667600 0.120151
-1000900 0.185303
-1334200 0.253382
-1667500 0.331956
-2000800 0.399671
-2334100 0.441368
-2667400 0.515117
-3000700 0.570733
-3334000 0.641394
-3667300 0.716672
-4000600 0.801108
-4333900 0.891152
-4667200 0.984463
-5000500 1.030872
-5333800 1.107730
-5667100 1.161548
-6000400 1.271633
-6333700 1.366231
-6667000 1.377119
-7000300 1.463278
-7333600 1.471678
-7666900 1.554606
-8000200 1.634424
-8333500 1.797473
-8666800 1.852463
-9000100 1.960743
-9333400 2.000188
-9666700 2.072302
-10000000 2.208677
