+++ resultados/tempo_seq_1
@@ -1,0 +1,31 @@
+1000 0,000248
+334300 0,063076
+667600 0,127294
+1000900 0,198593
+1334200 0,254572
+1667500 0,368523
+2000800 0,40432
+2334100 0,456859
+2667400 0,522386
+3000700 0,597894
+3334000 0,673851
+3667300 0,732865
+4000600 0,810892
+4333900 0,880156
+4667200 0,9589
+5000500 1,086666
+5333800 1,103852
+5667100 1,182942
+6000400 1,242411
+6333700 1,309282
+6667000 1,403294
+7000300 1,568971
+7333600 1,535038
+7666900 1,600342
+8000200 1,68086
+8333500 1,754961
+8666800 1,834468
+9000100 1,915818
+9333400 1,97759
+9666700 2,059755
+10000000 2,123877
