--- resultados/tempo_seq_comp
@@ -1,3 +1,0 @@
-10000000 2.208677
-10000000 2.278926
-10000000 2.125175
