--- resultados/tempo_seq_media
@@ -1,1 +1,0 @@
-1 2.204259
